Split the circular layout algo
